+++ entry.alpine.sh
@@ -23,6 +23,16 @@ fi
 
 export DBUS_SYSTEM_BUS_ADDRESS="unix:path=/mnt/root/run/dbus/system_bus_socket"
 
+# If DOCKER_ROOT isn't set then default it
+if [ -z "${DOCKER_ROOT}" ]; then
+	DOCKER_ROOT=/mnt/root/var/lib/rce
+fi
+# Mount the DOCKER_ROOT path equivalent in the container fs
+DOCKER_LIB_PATH=${DOCKER_ROOT#/mnt/root}
+if [ ! -d "${DOCKER_LIB_PATH}" ]; then
+	ln -s "${DOCKER_ROOT}" "${DOCKER_LIB_PATH}"
+fi
+
 /usr/bin/supervisord -c /etc/supervisor/supervisord.conf
 
 supervisorctl -c /etc/supervisor/supervisord.conf start resin-supervisor